moved all block-chain bindings to domain specific folders
