+++ packages/client/src/escrow/graphql/queries.ts
@@ -1,0 +1,16 @@
+export const getProtocolAndPlatformsFees = (
+    originServicePlatformId: string,
+    originValidatedProposalPlatformId: string,
+): string => `
+    {
+      protocols {
+        protocolEscrowFeeRate
+      }
+      servicePlatform: platform(id:${originServicePlatformId}){
+        originServiceFeeRate
+      }
+      proposalPlatform: platform(id:${originValidatedProposalPlatformId}){
+        originValidatedProposalFeeRate
+      }
+    }
+  `;
