+++ packages/client/src/graphql/queries.ts
@@ -80,59 +80,3 @@ export const getProposalsByService = (serviceId: string) => `
   }
 }
 `;
-
-export const getProposalById = (id: string) => `
-{
-  proposal(id: "${id}") {
-    id
-    seller {
-      id
-      handle
-      address
-      cid
-      rating
-      userStats {
-        numReceivedReviews
-      }
-    }
-    platform {
-      id
-    }
-    service {
-      id
-      platform {
-        id
-      }
-    }
-    cid
-    rateToken {
-      address
-    }
-    rateAmount
-    description {
-      about
-      video_url
-    }
-    status
-    expirationDate
-  }
-}
-`;
-
-export const getProtocolAndPlatformsFees = (
-  chainId: number,
-  originServicePlatformId: string,
-  originValidatedProposalPlatformId: string,
-): string => `
-  {
-    protocols {
-      protocolEscrowFeeRate
-    }
-    servicePlatform: platform(id:${originServicePlatformId}){
-      originServiceFeeRate
-    }
-    proposalPlatform: platform(id:${originValidatedProposalPlatformId}){
-      originValidatedProposalFeeRate
-    }
-  }
-`;
