+++ packages/client/src/services/enums/index.ts
@@ -1,0 +1,7 @@
+export enum ServiceStatusEnum {
+    Opened = "Opened",
+    Confirmed = "Confirmed",
+    Finished = "Finished",
+    Cancelled = "Cancelled",
+    Uncompleted = "Uncompleted",
+}
