+++ packages/client/src/services/graphql/queries.ts
@@ -1,0 +1,25 @@
+import { IProps } from "../types";
+
+export const getFilteredServiceDescriptionCondition = (params: IProps): string => {
+    let condition = ", where: {";
+    condition += params.serviceStatus
+        ? `service_: {status:"${params.serviceStatus}"}`
+        : "";
+    condition += params.buyerId ? `, buyer: "${params.buyerId}"` : "";
+    condition += params.sellerId ? `, seller: "${params.sellerId}"` : "";
+    condition += params.platformId ? `, platform: "${params.platformId}"` : "";
+    condition += "}";
+    return condition === ", where: {}" ? "" : condition;
+}
+
+export const getFilteredServiceCondition = (params: IProps): string => {
+    let condition = ", where: {";
+    condition += params.serviceStatus
+        ? `status: "${params.serviceStatus}"`
+        : "";
+    condition += params.buyerId ? `, buyer: "${params.buyerId}"` : "";
+    condition += params.sellerId ? `, seller: "${params.sellerId}"` : "";
+    condition += params.platformId ? `, platform: "${params.platformId}"` : "";
+    condition += "}";
+    return condition === ", where: {}" ? "" : condition;
+}
